+++ lib/models/users_provider.dart
@@ -1,0 +1,14 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+import 'user.dart';
+
+final usersProvider = StreamProvider.autoDispose<List<User>>((ref) {
+  return FirebaseFirestore.instance.collection('users').snapshots().map(
+        (snapshot) => snapshot.docs
+            .map(
+              (doc) => User.fromJsonDocument(doc.id, doc.data()),
+            )
+            .toList(),
+      );
+});