+++ lib/router/routes.g.dart
@@ -14,6 +14,7 @@ List<RouteBase> get $appRoutes => [
       $waitingForApprovalRoute,
       $settingsRoute,
       $gamesRoute,
+      $usersRoute,
     ];
 
 RouteBase get $homeRoute => GoRouteData.$route(
@@ -199,6 +200,28 @@ extension $GamesRouteExtension on GamesRoute {
 
   String get location => GoRouteData.$location(
         '/games',
+      );
+
+  void go(BuildContext context) => context.go(location);
+
+  Future<T?> push<T>(BuildContext context) => context.push<T>(location);
+
+  void pushReplacement(BuildContext context) =>
+      context.pushReplacement(location);
+
+  void replace(BuildContext context) => context.replace(location);
+}
+
+RouteBase get $usersRoute => GoRouteData.$route(
+      path: '/users',
+      factory: $UsersRouteExtension._fromState,
+    );
+
+extension $UsersRouteExtension on UsersRoute {
+  static UsersRoute _fromState(GoRouterState state) => const UsersRoute();
+
+  String get location => GoRouteData.$location(
+        '/users',
       );
 
   void go(BuildContext context) => context.go(location);