+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+- Adds a users screen.
+
 ## 0.7.0 [2023-10-08]
 
 - Adds a sign up flow until a dummy waiting for approval screen.